claude自動引当ツール/requirements.txt
flask>=3.1.0
flask-sqlalchemy>=3.1.0
python-dotenv>=1.0.0
requests>=2.32.0
gunicorn>=23.0.0
openpyxl>=3.1.0
APScheduler>=3.10.0
